feat: Add student creation endpoint and DTOs
Add a new endpoint for creating student records.
Improve data types for yearOfBirth in DTOs.
Enhance data fetching efficiency in entities.

+++ src/main/java/com/pos/point_of_sale/controller/student/dto/Mother.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public
 class Mother {
          private String name;
-         private Long yearOfBirth;
+         private Integer yearOfBirth;
          private String education;
          private String job;
          private String income;
